Feature: Brightness limit
This intercepts and overrides writes to $2100, limiting the brightness
written to a configurable value (8..15).
The FPGA feature bitfield is extended to 16 bits to accomodate the 4
additional bits for the brightness limit.
The menu itself does not apply the brightness limit on FPGA level but
"fakes" it by setting the brightness manually. The newly added menu
change hooks are used to reflect the brightness setting as soon as it is
changed during selection.

+++ src/smc.h
@@ -88,7 +88,7 @@ typedef struct __attribute__ ((__packed__)) _snes_romprops {
   uint8_t has_sa1;            /* SA-1 presence flag */
   uint8_t has_sdd1;           /* S-DD1 presence flag */
   uint8_t has_spc7110;        /* SPC7110 presence flag */
-  uint8_t fpga_features;      /* feature/peripheral enable bits */
+  uint16_t fpga_features;     /* feature/peripheral enable bits */
   uint16_t fpga_dspfeat;      /* DSP configuration bits */
   uint8_t region;             /* game region (derived from destination code) */
   uint32_t load_address;      /* where to load the ROM image */
